requirements.txt
streamlit
langchain-community
streamlit-option-menu
langchain
langchain-groq
python-dotenv
BeautifulSoup4
langchain_cohere
chromadb
duckduckgo-search
langchain-google-genai
pdfminer.six